requirements.txt
pillow==11.2.1
pytesseract==0.3.13
deep-translator==1.11.4